rules/requirements.txt
python-Levenshtein==0.12.2
transformers==3.5.0
